Feature/create domain layer

+++ app/src/main/java/com/alicimsamil/harmonyhub/core/common/model/TrackDBModel.kt
@@ -48,7 +48,7 @@ data class TrackDBModel(
     @ColumnInfo(name = "trackRentalPrice")
     val trackRentalPrice: Double?,
     @ColumnInfo(name = "trackTimeMillis")
-    val trackTimeMillis: Int?,
+    val trackTimeMillis: Long?,
     @ColumnInfo(name = "wrapperType")
     val wrapperType: String?,
     @PrimaryKey(autoGenerate = true)
